+++ GitUI/GitUI.csproj
@@ -41,6 +41,9 @@
       <LastGenOutput>Images.Designer.cs</LastGenOutput>
     </EmbeddedResource>
     <EmbeddedResource Update="Properties\Resources.resx" Generator="ResXFileCodeGenerator" LastGenOutput="Resources.Designer.cs" />
+    <Compile Update="CommandsDialogs\Menus\StartToolStripMenuItem.cs">
+      <SubType>Component</SubType>
+    </Compile>
     <Compile Update="UserControls\FilterToolBar.Designer.cs">
       <SubType>Component</SubType>
     </Compile>
